Int to double, and sizing componemt

+++ Sources/Infrastructure/Domain/Models/Product.cs
@@ -18,9 +18,9 @@ namespace Domain.Models
         [Required]
         public string Name { get; set; } = default!;
         public string? Description { get; set; }
-        public int? Height { get; set; }
-        public int? TopDiameter { get; set; }
-        public int? BottomDiameter { get; set; }
+        public double? Height { get; set; }
+        public double? TopDiameter { get; set; }
+        public double? BottomDiameter { get; set; }
         public string? DesignInstruction { get; set; }
         public int? Status { get; set; }
         public int MarginRate { get; set; }
